+++ tests/ami/scripts/fresh_install_deb.sh
