+++ tests/ami/scripts/fresh_install_rpm.sh
