+++ tests/ami/scripts/upgrade_install_rpm.sh
